System/Backend/requirements.txt
alembic==1.12.1
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.1.2
boto3==1.26.101
botocore==1.29.165
certifi==2025.4.26
Charm-Crypto==0.50
charset-normalizer==3.4.2
click==8.1.8
dotenv==0.9.9
exceptiongroup==1.2.2
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
hypothesis==6.131.15
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.0
packaging==24.0
psycopg2-binary==2.9.9
pycrypto==2.6.1
pycryptodome==3.22.0
PyJWT==2.8.0
pyparsing==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
requests==2.31.0
s3transfer==0.6.2
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
tabulate==0.9.0
termcolor==2.3.0
typing_extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.2.3
zipp==3.15.0
